Add 3 more tests to fit requirements according to the CR
