+++ app/views/posts/show.html.erb
@@ -1,6 +1,6 @@
 <section class="full-post-section">
   <div class="full-post border-black">
-    <h2>Post # <%= @post.id %> by <%= @post.author.name %></h2>
+     <h2>Post # <%= @post.id %>: <%= @post.title %> by <%= @post.author.name %></h2>
     <p class="post-likes-comments">Comments: <%= @post.comments_counter %>, Likes: <%= @post.likes_counter %></p>
     <p><%= @post.text %></p>
         <div class="like-comment-btn">
