+++ app/views/shared/_post.html.erb
@@ -1,5 +1,5 @@
 <div class="user-post border-black">
-  <h2>Post #<%= post.id %></h2>
+  <h2>Post #<%= post.id %></h2> : <%= post.title %></h2>
   <p><%= truncate(post.text, length: 150, omission: "...") %></p>
   <div class="like-comment">
     <p>Comments: <%= post.comments_counter %>, Likes: <%= post.likes_counter %></p>
